fix full_cache table  insert update remove SortForm system call,  and start refactoring pre release

+++ src/main/java/org/folio/marccat/business/cataloguing/bibliographic/BibliographicCatalog.java
@@ -25,10 +25,6 @@ import static org.folio.marccat.util.F.isNotNull;
 
 /**
  * Bibliographic implementation of {@link Catalog} interface.
- *
- * @author paulm
- * @author janick
- * @author cchiama
  */
 public class BibliographicCatalog extends Catalog {
 
@@ -415,7 +411,7 @@ public class BibliographicCatalog extends Catalog {
             tag = new PublisherManager();
             tag.setCorrelationValues(correlationValues);
           } else {
-            tag = new PublisherManager(item.getAmicusNumber().intValue(), item.getUserView());
+            tag = new PublisherManager(item.getAmicusNumber(), item.getUserView());
             tag.setCorrelationValues(correlationValues);
           }
           tag = (Tag) setItemIfNecessary(item, tag);
